+++ Code/EasyPacketDLL.cs
@@ -3,37 +3,45 @@
  *  DavidFDev
  */
 
-using System;
 using System.IO;
 using EasyPacketsLib.Internals;
 using Terraria.ModLoader;
 
 namespace EasyPacketsLib;
 
+/// <summary>
+///     Methods for using the library as a DLL reference instead of as a mod reference.
+/// </summary>
 public static class EasyPacketDLL
 {
     #region Static Methods
 
+    /// <summary>
+    ///     Register easy packets and handlers of the provided mod.
+    ///     If registering multiple mods, ensure the order is the same across clients.
+    ///     <example>
+    ///         <code>public override void Load() => EasyPacketDLL.RegisterMod(this);</code>
+    ///     </example>
+    /// </summary>
+    /// <param name="mod">Mod to load the types from.</param>
     public static void RegisterMod(Mod mod)
     {
-        // NOTE in XML: if registering multiple mods for some reason, order is important
-        
         // Ensure this mod is using a DLL reference only
         if (ModContent.GetInstance<EasyPacketsLibMod>() != null)
         {
-            mod.Logger.Warn($"IGNORE REGISTER MOD FOR: " + mod.Name);
-            return;
-        }
-
-        if (mod.Side is not ModSide.Both)
-        {
-            mod.Logger.Warn($"MUST BE BOTH: " + mod.Name);
+            mod.Logger.Error("Failed to manually register easy packets for mod because it is not a DLL reference");
             return;
         }
 
         EasyPacketLoader.RegisterMod(mod);
     }
 
+    /// <summary>
+    ///     Handle incoming easy packets.
+    ///     <example>
+    ///         <code>public override void HandlePacket(BinaryReader reader, int whoAmI) => EasyPacketDLL.HandlePacket(reader, whoAmI);</code>
+    ///     </example>
+    /// </summary>
     public static void HandlePacket(BinaryReader reader, int whoAmI)
     {
         // Ensure this mod is using a DLL reference only